+++ packages/nadle/src/core/nadle.ts
@@ -10,8 +10,10 @@ import { TaskRegistry } from "./registration/task-registry.js";
 import { OptionsResolver } from "./options/options-resolver.js";
 import { type State, type ExecutionContext } from "./context.js";
 import { ExecutionTracker } from "./models/execution-tracker.js";
+import { RootWorkspace } from "./models/project/root-workspace.js";
 import { DefaultLogger } from "./interfaces/defaults/default-logger.js";
 import { FileOptionRegistry } from "./registration/file-option-registry.js";
+import { DefaultFileReader } from "./interfaces/defaults/default-file-reader.js";
 import { type NadleCLIOptions, type NadleResolvedOptions } from "./options/types.js";
 
 export class Nadle implements ExecutionContext {
@@ -37,6 +39,34 @@ export class Nadle implements ExecutionContext {
 		await this.eventEmitter.onInitialize();
 
 		return this;
+	}
+
+	public async initForWorker(resolvedOptions: NadleResolvedOptions): Promise<this> {
+		this.#options = resolvedOptions;
+		this.logger.configure(resolvedOptions);
+		this.logger.debug("Worker init: loading config files (skipping project resolution)");
+
+		await runWithInstance(this, () => this.loadConfigFiles(resolvedOptions));
+		this.taskRegistry.configure(resolvedOptions.project);
+
+		return this;
+	}
+
+	private async loadConfigFiles(resolvedOptions: NadleResolvedOptions) {
+		const { project } = resolvedOptions;
+		const fileReader = new DefaultFileReader();
+
+		this.taskRegistry.onConfigureWorkspace(RootWorkspace.ID);
+		this.fileOptionRegistry.onConfigureWorkspace(RootWorkspace.ID);
+		await fileReader.read(project.rootWorkspace.configFilePath);
+
+		for (const workspace of project.workspaces) {
+			if (workspace.configFilePath) {
+				this.taskRegistry.onConfigureWorkspace(workspace.id);
+				this.fileOptionRegistry.onConfigureWorkspace(workspace.id);
+				await fileReader.read(workspace.configFilePath);
+			}
+		}
 	}
 
 	public async execute() {